All face textures fixed. Stamina bar regen is nearly finished.

+++ Assets/Scripts/Entities/Player/UI/UIStatBar.cs
@@ -13,11 +13,11 @@ public class UIStatBar : MonoBehaviour
         slider = GetComponent<Slider>();
     }
 
-    public virtual void SetStat(int newValue) {
+    public virtual void SetStat(float newValue) {
         slider.value = newValue;
     }
 
-    public virtual void SetMaxStat(int maxValue) {
+    public virtual void SetMaxStat(float maxValue) {
         slider.maxValue = maxValue;
         slider.value = maxValue;
     }
